Added second main class 'Game2' which renders differently to see which version we should use. Added player movement using WASD. Adding collision detection between fish.

+++ src/main/EnemyFish.java
@@ -2,7 +2,7 @@ package main;
 
 public class EnemyFish extends Entity {
 	
-	public EnemyFish(double movespeed, boolean isAlive, Sprite sprite) {
+	public EnemyFish(int movespeed, boolean isAlive, Sprite sprite) {
 		super(movespeed, isAlive, sprite);
 	}
 }
