+++ src/main/PlayerFish.java
@@ -2,7 +2,7 @@ package main;
 
 public class PlayerFish extends Entity {
 	
-	public PlayerFish(double movespeed, boolean isAlive, Sprite sprite) {
+	public PlayerFish(int movespeed, boolean isAlive, Sprite sprite) {
 		super(movespeed, isAlive, sprite);
 	}
